Refactor file system and import path for base/core.py
The base/core.py file was moved to core.py which required refactoring of import paths in encoding.py. This allows for more intuitive file system structure and easier navigation through the module.
