+++ src/mdio/schemas/base/encoding.py
@@ -3,9 +3,9 @@
 
 from pydantic import Field
 
-from mdio.schemas.base.core import StrictCamelBaseModel
 from mdio.schemas.chunk_grid import RectilinearChunkGrid
 from mdio.schemas.chunk_grid import RegularChunkGrid
+from mdio.schemas.core import StrictCamelBaseModel
 
 
 class ChunkGridMetadata(StrictCamelBaseModel):
