+++ src/mdio/schemas/core.py
